Add utils with some helper functions and move EIP-55 check to parsers
