+++ packages/siwe/lib/client.ts
@@ -1,7 +1,6 @@
 import { ABNFParsedMessage, RegExpParsedMessage } from '@spruceid/siwe-parser';
-import { randomStringForEntropy } from '@stablelib/random';
 // TODO: Figure out how to get types from this lib:
-import { Contract, ethers, utils } from 'ethers';
+import { ethers } from 'ethers';
 
 /**
  * Possible message error types.
@@ -103,10 +102,6 @@ export class SiweMessage {
 				this.chainId = parseInt(this.chainId)
 			}
 		}
-
-		if (this.address !== utils.getAddress(this.address)) {
-			throw new Error('Address not in EIP 55 format');
-		}
 	}
 
 	/**
@@ -130,7 +125,7 @@ export class SiweMessage {
 	toMessage(): string {
 		const header = `${this.domain} wants you to sign in with your Ethereum account:`;
 		const uriField = `URI: ${this.uri}`;
-		let prefix = [header, utils.getAddress(this.address)].join('\n');
+		let prefix = [header, this.address].join('\n');
 		const versionField = `Version: ${this.version}`;
 
 		if (!this.nonce) {
@@ -296,49 +291,3 @@ export class SiweMessage {
 		});
 	}
 }
-
-/**
- * This method calls the EIP-1271 method for Smart Contract wallets
- * @param message The EIP-4361 parsed message
- * @param provider Web3 provider able to perform a contract check (Web3/EthersJS).
- * @returns {Promise<boolean>} Checks for the smart contract (if it exists) if
- * the signature is valid for given address.
- */
-export const checkContractWalletSignature = async (
-	message: SiweMessage,
-	signature: string,
-	provider?: any
-): Promise<boolean> => {
-	if (!provider) {
-		return false;
-	}
-
-	const abi = [
-		'function isValidSignature(bytes32 _message, bytes _signature) public view returns (bool)',
-	];
-	try {
-		const walletContract = new Contract(message.address, abi, provider);
-		const hashMessage = utils.hashMessage(message.signMessage());
-		return await walletContract.isValidSignature(
-			hashMessage,
-			signature,
-		);
-	} catch (e) {
-		throw e;
-	}
-};
-
-/**
- * This method leverages a native CSPRNG with support for both browser and Node.js
- * environments in order generate a cryptographically secure nonce for use in the
- * SiweMessage in order to prevent replay attacks.
- *
- * 96 bits has been chosen as a number to sufficiently balance size and security considerations
- * relative to the lifespan of it's usage.
- *
- * @returns cryptographically generated random nonce with 96 bits of entropy encoded with
- * an alphanumeric character set.
- */
-export const generateNonce = (): string => {
-	return randomStringForEntropy(96);
-};
